Adicionado Pattern Facade e corrigido bug loop de estadia com problema.

+++ src/main/java/br/com/code/pethotel/repository/EstadiaRepository.java
@@ -3,5 +3,5 @@ package br.com.code.pethotel.repository;
 import br.com.code.pethotel.model.Estadia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadiaRepository extends JpaRepository<Estadia, Integer> {
+public interface EstadiaRepository extends JpaRepository<Estadia, Long> {
 }
