+++ src/main/java/br/com/code/pethotel/service/EstadiaService.java
@@ -8,8 +8,8 @@ import java.util.Optional;
 public interface EstadiaService {
 
 	Estadia createEstadiaPet(Estadia estadia);
-	Estadia updateEstadiaPet(Estadia estadia, int id);
-	Optional<Estadia> findEstadiaPetById(int id);
+	Estadia updateEstadiaPet(Estadia estadia, Long id);
+	Optional<Estadia> findEstadiaPetById(Long id);
 	List<Estadia> listAllEstadias();
 
 }
